Refactor into a few packages
